Add files via upload
Traditional paper-based queue management systems have been accepted as the industry standard for far too long and have been adopted by businesses of all sizes. While the world has evolved and adopted new technologies better suited to today's needs, queueing has largely remained unchanged. Current methods not only generate a huge amount of thermal paper waste that is neither recyclable nor compostable, but they also hardly address operational efficiency and customer experience. People waste hours waiting in crowded places when all they need to do is be present when it's their turn.

Waqty was created as a future-proof queue management solution with flexibility in mind. We make it possible for people to reserve or schedule their position in line regardless of where they are, and by collaborating closely with companies, we will make sure that there is no shortage of strong tools to easily and transparently manage and monitor locations. If there’s a queue, we streamline it, as we are neither industry specific, nor tightly coupled with proprietary hardware. Therefore, our solution will fit in well with your company, whether you're a doctor managing a clinic or a barber managing a barber shop trying to better serve your consumers.